# #0: Remove bit fields from edm packet header