### Ticket
Link to Github Issue

### Problem description
Bit fields in packet header causes additional instructions like shifting.

### What's changed
Remove bit fields from packet header.
Unicast: 17.47 GB/s -> 17.74 GB/s
Mcast: 14.02 GB/s -> 14.12 GB/s

### Checklist
- [ ] [All post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml) CI passes
- [ ] [Blackhole Post commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml) CI passes (if applicable)
- [ ] [Model regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml) CI passes (if applicable)
- [ ] [Device performance regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml) CI passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) CI passes (if applicable)
- [x] New/Existing tests provide coverage for changes
T3K nightly: https://github.com/tenstorrent/tt-metal/actions/runs/13547469910
Microbenchmarks: https://github.com/tenstorrent/tt-metal/actions/runs/13544943235